refactor: remove wearables v2 flag (#2451)
