+++ kernel/packages/config/index.ts
@@ -135,8 +135,6 @@ export const PIN_CATALYST = qs.CATALYST ? addHttpsIfNoProtocolIsSet(qs.CATALYST)
 
 export const FORCE_RENDERING_STYLE = qs.FORCE_RENDERING_STYLE
 
-export const TEST_WEARABLES_OVERRIDE = location.search.includes('TEST_WEARABLES')
-
 const META_CONFIG_URL = qs.META_CONFIG_URL
 
 export namespace commConfigurations {
@@ -242,7 +240,6 @@ export function getExclusiveServer() {
   return 'https://wearable-api.decentraland.org/v2/collections'
 }
 
-export const ALL_WEARABLES = location.search.includes('ALL_WEARABLES') && getDefaultTLD() !== 'org'
 export const WITH_FIXED_COLLECTIONS = qs.WITH_COLLECTIONS && getDefaultTLD() !== 'org' ? qs.WITH_COLLECTIONS : undefined
 export const WEARABLE_API_DOMAIN = qs.WEARABLE_API_DOMAIN || 'wearable-api.decentraland.org'
 export const WEARABLE_API_PATH_PREFIX = qs.WEARABLE_API_PATH_PREFIX || 'v2'
