+++ kernel/packages/shared/catalogs/selectors.ts
@@ -8,17 +8,8 @@ export const getPlatformCatalog = (store: RootCatalogState): Record<WearableId, 
     ? store.catalogs.catalogs['base-avatars'].data!
     : null
 
-export const getExclusiveCatalog = (store: RootCatalogState): Record<WearableId, PartialWearableV2> | null =>
-  store.catalogs.catalogs &&
-  store.catalogs.catalogs['base-exclusive'] &&
-  store.catalogs.catalogs['base-exclusive'].status === 'ok'
-    ? store.catalogs.catalogs['base-exclusive'].data!
-    : null
-
 export const baseCatalogsLoaded = (store: RootCatalogState) =>
   store.catalogs &&
   store.catalogs.catalogs &&
   store.catalogs.catalogs['base-avatars'] &&
-  store.catalogs.catalogs['base-avatars'].status === 'ok' &&
-  store.catalogs.catalogs['base-exclusive'] &&
-  store.catalogs.catalogs['base-exclusive'].status === 'ok'
+  store.catalogs.catalogs['base-avatars'].status === 'ok'
