+++ kernel/test/unit/catalog.saga.test.tsx
@@ -1,11 +1,9 @@
 import { expectSaga } from 'redux-saga-test-plan'
 import { call, select } from 'redux-saga/effects'
-import { BASE_AVATARS_COLLECTION_ID, fetchInventoryItemsByAddress, handleWearablesFailure, handleWearablesRequest, handleWearablesSuccess, informRequestFailure, sendWearablesCatalog, WRONG_FILTERS_ERROR } from 'shared/catalogs/sagas'
+import { handleWearablesFailure, handleWearablesRequest, handleWearablesSuccess, informRequestFailure, sendWearablesCatalog, WRONG_FILTERS_ERROR } from 'shared/catalogs/sagas'
 import { wearablesFailure, wearablesRequest, wearablesSuccess } from 'shared/catalogs/actions'
-import { baseCatalogsLoaded, getExclusiveCatalog, getPlatformCatalog } from 'shared/catalogs/selectors'
+import { baseCatalogsLoaded, getPlatformCatalog } from 'shared/catalogs/selectors'
 import { ensureRenderer } from 'shared/renderer/sagas'
-import { throwError } from 'redux-saga-test-plan/providers'
-import { getFetchContentServer } from 'shared/dao/selectors'
 
 const serverUrl = 'https://server.com'
 const wearableId1 = 'WearableId1'
@@ -16,46 +14,6 @@ const context = 'someContext'
 
 
 describe('Wearables Saga', () => {
-
-  it('When a wearable is requested by id, then it is returned successfully', () => {
-    return expectSaga(handleWearablesRequest, wearablesRequest({ wearableIds: [wearableId1] }, context))
-      .put(wearablesSuccess([wearable1], context))
-      .provide([
-        [select(getFetchContentServer), serverUrl],
-        [select(baseCatalogsLoaded), true],
-        [select(getPlatformCatalog), {}],
-        [select(getExclusiveCatalog), { [wearableId1]: wearable1 }],
-      ])
-      .run()
-  })
-
-  it('When all owned wearables are requested, then they are returned successfully', () => {
-    const wearables = [wearable1]
-    return expectSaga(handleWearablesRequest, wearablesRequest({ ownedByUser: userId }, context))
-      .put(wearablesSuccess(wearables, context))
-      .provide([
-        [select(getFetchContentServer), serverUrl],
-        [select(baseCatalogsLoaded), true],
-        [select(getPlatformCatalog), {}],
-        [select(getExclusiveCatalog), { [wearableId1]: wearable1 }],
-        [call(fetchInventoryItemsByAddress, userId), Promise.resolve([wearableId1])],
-      ])
-      .run()
-  })
-
-  it('When base avatars are requested, then they are returned successfully', () => {
-    const baseWearables = [wearable1]
-    return expectSaga(handleWearablesRequest, wearablesRequest({ collectionIds: [BASE_AVATARS_COLLECTION_ID] }, context))
-      .put(wearablesSuccess(baseWearables, context))
-      .provide([
-        [select(getFetchContentServer), serverUrl],
-        [select(baseCatalogsLoaded), true],
-        [select(getPlatformCatalog), { [wearableId1]: wearable1 }],
-        [select(getExclusiveCatalog), {}],
-      ])
-      .run()
-  })
-
 
   it('When wearables fetch is successful, then it is sent to the renderer with the same context', () => {
     const wearables = [wearable1]
@@ -68,28 +26,12 @@ describe('Wearables Saga', () => {
       .run()
   })
 
-  it('When wearables fetch fails to load, then the request fails', () => {
-    const errorMessage = 'Something failed'
-    const error = new Error(errorMessage)
-    return expectSaga(handleWearablesRequest, wearablesRequest({ ownedByUser: userId }, context))
-      .put(wearablesFailure(context, errorMessage))
-      .provide([
-        [select(getFetchContentServer), serverUrl],
-        [select(baseCatalogsLoaded), true],
-        [select(getPlatformCatalog), {}],
-        [select(getExclusiveCatalog), {}],
-        [call(fetchInventoryItemsByAddress, userId), throwError(error)],
-      ])
-      .run()
-  })
-
   it('When more than one filter is set, then the request fails', () => {
     return expectSaga(handleWearablesRequest, wearablesRequest({ wearableIds: ['some-id'], ownedByUser: userId }, context))
       .put(wearablesFailure(context, WRONG_FILTERS_ERROR))
       .provide([
         [select(baseCatalogsLoaded), true],
         [select(getPlatformCatalog), {}],
-        [select(getExclusiveCatalog), {}],
       ])
       .run()
   })
